Finally Working

+++ process-video.py
@@ -1,3 +1,4 @@
+from math import ceil
 import subprocess
 import sys
 from pytube import YouTube
@@ -7,8 +8,18 @@ from openai import OpenAI
 from dotenv import load_dotenv
 import json
 from moviepy.video.io.VideoFileClip import VideoFileClip
+import numpy as np
+import pickle
+import mimetypes
+from google.auth.transport.requests import Request
+from google.oauth2.credentials import Credentials
+from google_auth_oauthlib.flow import InstalledAppFlow
+from googleapiclient.discovery import build
+from googleapiclient.http import MediaFileUpload
 
 load_dotenv()
+
+SCOPES = ["https://www.googleapis.com/auth/drive.file"]
 
 
 def main():
@@ -21,71 +32,97 @@ def main():
 
     # Download the video from the url
     print("Downloading video...")
-    raw_video = download_video(youtube_url, 'Raw')
+    raw_video = download_video(youtube_url, "Raw")
 
     # Get the Transcript
     print("Getting transcript...")
     transcript = get_transcript(youtube_url)
-    transcript_text_list = [entry['text'] for entry in transcript]
-    transcript_text = ' '.join(transcript_text_list)
+    transcript_text_list = [entry["text"] for entry in transcript]
+    transcript_text = " ".join(transcript_text_list)
 
     # Check video duration
     video = VideoFileClip(raw_video)
-    duration = video.duration
 
-    if duration > 3600:  # If the video is longer than an hour
-        print("Video is longer than an hour. Splitting the video...")
-        mid_point = duration / 2
-
-        # Split the video into two halves
-        first_half = video.subclip(0, mid_point)
-        second_half = video.subclip(mid_point, duration)
-
-        # Split the transcript into two halves
-        first_half_transcript = [entry for entry in transcript if entry['start'] < mid_point]
-        second_half_transcript = [entry for entry in transcript if entry['start'] >= mid_point]
-
-        video_name = os.path.basename(raw_video).split('.')[0]
-
-        # Process the first half
-        print("Processing the first half of the video...")
-        first_half_path = 'Raw/' + video_name + '_first_half.mp4'
-        first_half.write_videofile(first_half_path)
-        first_half_transcript = [entry for entry in transcript if entry['start'] < mid_point]
-        first_half_transcript_text_list = [entry['text'] for entry in first_half_transcript]
-        first_half_transcript_text = ' '.join(first_half_transcript_text_list)
-        first_half_interesting_parts = get_interesting_parts(first_half_transcript_text)
-        first_half_timestamps = extract_timestamps(first_half_interesting_parts, first_half_transcript)
-        crop_video(first_half_path, first_half_timestamps, 'Cropped')
-
-        # Process the second half
-        print("Processing the second half of the video...")
-        second_half_path = 'Raw/' + video_name + '_second_half.mp4'
-        second_half.write_videofile(second_half_path)
-        second_half_transcript = [entry for entry in transcript if entry['start'] >= mid_point]
-        second_half_transcript_text_list = [entry['text'] for entry in second_half_transcript]
-        second_half_transcript_text = ' '.join(second_half_transcript_text_list)
-        second_half_interesting_parts = get_interesting_parts(second_half_transcript_text)
-        second_half_timestamps = extract_timestamps(second_half_interesting_parts, second_half_transcript)
-        crop_video(second_half_path, second_half_timestamps, 'Cropped')
-
+    # Get the interesting parts of the video
+    print("Getting interesting parts...")
+    split_coefficient = 2400
+    if video.duration > split_coefficient:
+        split = ceil(video.duration / split_coefficient)
     else:
-        # Get the interesting parts of the video
-        print("Getting interesting parts...")
-        interesting_parts = get_interesting_parts(transcript_text)
+        split = 1
 
-        # Extract timestamps for the most interesting parts
-        print("Extracting timestamps...")
-        timestamps = extract_timestamps(interesting_parts, transcript)
+    interesting_parts = get_interesting_parts(transcript_text, split)
 
-        print(timestamps)
+    # Extract timestamps for the most interesting parts
+    print("Extracting timestamps...")
+    timestamps = extract_timestamps(interesting_parts, transcript)
 
-        # Crop the video to the interesting parts
-        print("Cropping video...")
-        crop_video(raw_video, timestamps, 'Cropped')
+    print(timestamps)
+
+    # Crop the video to the interesting parts
+    print("Cropping video...")
+    crop_video(raw_video, timestamps, "Cropped")
+
+    print("Uploading videos to Google Drive...")
+    upload_videos_to_drive()
+
+    print("Cleaning up...")
+    clear_folder("Cropped")
 
     print("Processing completed.")
 
+
+def clear_folder(folder_path):
+    for filename in os.listdir(folder_path):
+        file_path = os.path.join(folder_path, filename)
+        os.remove(file_path)
+
+
+def upload_videos_to_drive():
+    drive_service = authenticate_drive_api()
+    folder_path = "Cropped"
+    folder_id = os.getenv("DRIVE_FOLDER_ID")
+
+    for filename in os.listdir(folder_path):
+        file_path = os.path.join(folder_path, filename)
+        if os.path.isfile(file_path) and filename.endswith(
+            ("mp4", "avi", "mov", "mkv")
+        ):
+            file_metadata = {"name": filename, "parents": [folder_id]}
+            mime_type, _ = mimetypes.guess_type(file_path)
+            media = MediaFileUpload(file_path, mimetype=mime_type)
+
+            file = (
+                drive_service.files()
+                .create(body=file_metadata, media_body=media, fields="id")
+                .execute()
+            )
+            print(f'File ID: {file.get("id")} uploaded successfully!')
+
+
+def authenticate_drive_api():
+    """Shows basic usage of the Drive v3 API.
+    Prints the names and ids of the first 10 files the user has access to.
+    """
+    creds = None
+    # The file token.pickle stores the user's access and refresh tokens, and is
+    # created automatically when the authorization flow completes for the first
+    # time.
+    if os.path.exists("token.pickle"):
+        with open("token.pickle", "rb") as token:
+            creds = pickle.load(token)
+    # If there are no (valid) credentials available, let the user log in.
+    if not creds or not creds.valid:
+        if creds and creds.expired and creds.refresh_token:
+            creds.refresh(Request())
+        else:
+            flow = InstalledAppFlow.from_client_secrets_file("credentials.json", SCOPES)
+            creds = flow.run_local_server(port=0)
+        # Save the credentials for the next run
+        with open("token.pickle", "wb") as token:
+            pickle.dump(creds, token)
+
+    return build("drive", "v3", credentials=creds)
 
 
 def crop_video(input_file, timestamps, output_path):
@@ -98,23 +135,39 @@ def crop_video(input_file, timestamps, output_path):
     # For each timestamp, create a new clip from 60 seconds before to 60 seconds after the timestamp
     for i, timestamp in enumerate(timestamps[:10]):  # Limit to the first 10 timestamps
         start = max(0, timestamp - 60)  # Ensure start is not negative
-        end = timestamp + 60
-        clip = video.subclip(start, end)
+        end = min(
+            video.duration, timestamp + 60
+        )  # Set end to the minimum of video duration or timestamp + 60
+        if end - start >= 1:
+            clip = video.subclip(start, end)
 
-        # Write the clip to a file
-        clip.write_videofile(f"{output_path}/{filename}_{i}.mp4")
+            # Write the clip to a file
+            clip.write_videofile(
+                f"{output_path}/{filename}_{i}.mp4", codec="h264_nvenc"
+            )
 
-        clip.close()
+            clip.close()
 
     video.close()
 
 
 def extract_timestamps(interesting_parts, transcript):
     timestamps = []
-    for entry in transcript:
+    for i in range(len(transcript)):
+        text = transcript[i]["text"]
+        j = i + 1
+
+        # Concatenate the text of consecutive entries until the text contains more than 5 words
+        while len(text.split()) < 3 and j < len(transcript):
+            text += " " + transcript[j]["text"]
+            j += 1
+
         for sentence in interesting_parts:
-            if entry['text'] in sentence:
-                timestamps.append(entry['start'])
+            if text in sentence:
+                print("sentence: ", sentence)
+                print("text: ", text)
+                if transcript[i]["start"] not in timestamps:
+                    timestamps.append(transcript[i]["start"])
                 interesting_parts.remove(sentence)
 
     return timestamps
@@ -122,13 +175,17 @@ def extract_timestamps(interesting_parts, transcript):
 
 def remove_silence(input_file, output_file=None):
     if output_file is None:
-        output_file = input_file.rsplit('.', 1)[0] + '_ALTERED.mp4'
+        output_file = input_file.rsplit(".", 1)[0] + "_ALTERED.mp4"
 
     command = [
-        'auto-editor', input_file,
-        '--edit', 'audio:threshold=0.04,stream=all',  # Cuts out sections where the audio volume is below 4% in all streams
-        '--margin', '0.2sec',  # Adds 0.2 seconds of padding around "loud" sections
-        '--output', output_file
+        "auto-editor",
+        input_file,
+        "--edit",
+        "audio:threshold=0.04,stream=all",  # Cuts out sections where the audio volume is below 4% in all streams
+        "--margin",
+        "0.2sec",  # Adds 0.2 seconds of padding around "loud" sections
+        "--output",
+        output_file,
     ]
 
     subprocess.run(command, check=True)
@@ -138,8 +195,13 @@ def remove_silence(input_file, output_file=None):
 def download_video(youtube_url, output_path):
     yt = YouTube(youtube_url)
 
-    video_stream = yt.streams.filter(progressive=False, file_extension='mp4').order_by('resolution').desc().first()
-    audio_stream = yt.streams.filter(only_audio=True).order_by('abr').desc().first()
+    video_stream = (
+        yt.streams.filter(progressive=False, file_extension="mp4")
+        .order_by("resolution")
+        .desc()
+        .first()
+    )
+    audio_stream = yt.streams.filter(only_audio=True).order_by("abr").desc().first()
 
     video_filename = video_stream.default_filename
     audio_filename = "audio_" + video_filename
@@ -147,19 +209,29 @@ def download_video(youtube_url, output_path):
     video_path = os.path.join(output_path, video_filename)
     audio_path = os.path.join(output_path, audio_filename)
 
-    video_stream.download(filename=video_filename, output_path=output_path)
-    audio_stream.download(filename=audio_filename, output_path=output_path)
-
-    # Combine video and audio
     output_file = os.path.join(output_path, "final_" + video_filename)
-    command = [
-        'ffmpeg', '-i', video_path, '-i', audio_path, '-c:v', 'copy', '-c:a', 'aac', output_file
-    ]
-    subprocess.run(command, check=True)
 
-    # Clean up and delete the uncombined files
-    os.remove(video_path)
-    os.remove(audio_path)
+    if not os.path.exists(output_file):
+        video_stream.download(filename=video_filename, output_path=output_path)
+        audio_stream.download(filename=audio_filename, output_path=output_path)
+
+        command = [
+            "ffmpeg",
+            "-i",
+            video_path,
+            "-i",
+            audio_path,
+            "-c:v",
+            "copy",
+            "-c:a",
+            "aac",
+            output_file,
+        ]
+        subprocess.run(command, check=True)
+
+        # Clean up and delete the uncombined files
+        os.remove(video_path)
+        os.remove(audio_path)
 
     return output_file
 
@@ -170,29 +242,45 @@ def get_transcript(video_url):
     return transcript
 
 
-def get_interesting_parts(transcript):
+def get_interesting_parts(transcript, split):
     # Concatenate transcript entries into a single text
-    client = OpenAI(api_key=os.getenv('OPENAI_API_KEY'))
+    client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))
 
-    # Use OpenAI API to summarize and extract key points
-    response = client.chat.completions.create(
-        model="gpt-3.5-turbo",
-        response_format={ "type": "json_object" },
-        messages=[
-            {"role": "system", "content": "You are a helpful assistant that finds the most interesting parts of a video transcript"},
-            {"role": "user", "content": "The following is the transcript of the video as a list of transcript entires. Identify the 10 most interesting sentences in the transcript. \
-                                        You should return a json object with the key 'interesting_parts' and a list of only the 10 most interesting sentences. Here is the transcript: " + str(transcript)}
-        ]
-    )
+    # Calculate the size of each part
+    part_size = ceil(len(transcript) / split)
+    print("split: ", split)
+    print("part_size: ", part_size)
 
-    # Print token summary of usgae
-    print("Usage:")
-    print(response.usage)
+    interesting_parts = []
 
-    summary = json.loads(response.choices[0].message.content)
+    # Use OpenAI API to summarize and extract key points for each part
+    for i in range(0, len(transcript), part_size):
+        part = transcript[i : min(i + part_size, len(transcript))]
+        response = client.chat.completions.create(
+            model="gpt-3.5-turbo",
+            response_format={"type": "json_object"},
+            messages=[
+                {
+                    "role": "system",
+                    "content": "You are a helpful assistant that finds the most interesting sentences of a video transcript",
+                },
+                {
+                    "role": "user",
+                    "content": "The following is the transcript of the video. Identify the 10 most interesting sentences in the transcript that would go viral if turned into a clip. \
+                                            You should return a json object with the key 'interesting_parts' and a list of only the 10 most interesting sentences. Here is the transcript: "
+                    + part,
+                },
+            ],
+        )
 
-    # Parse the JSON and get the list of entries
-    interesting_parts = summary['interesting_parts']
+        # Print token summary of usage
+        print("Usage:")
+        print(response.usage)
+
+        summary = json.loads(response.choices[0].message.content)
+
+        # Parse the JSON and get the list of entries
+        interesting_parts.extend(summary["interesting_parts"])
 
     return interesting_parts
